fix: node sort indices overflow past 65535 nodes (Uint16 -> Uint32)
_sortIndices held node indices up to maxNodes (100k default) in a
Uint16Array: past 65535 the values wrapped, duplicating indices. The
camera-distance sort then drew some nodes twice and dropped others
(34,464 corrupted entries at 100k nodes), and instanceId picking
(getDataByInstanceId) resolved to the wrong node.

Uint32Array matches the index arrays the link renderer already uses.
Cost: +195KB against the ~11MB the renderer already allocates; sort CPU
is unchanged (comparator-dominated, index width irrelevant).

+++ src/utils/InstancedNodeRenderer.js
@@ -41,8 +41,11 @@ export default class InstancedNodeRenderer {
     this.colorArray = new Float32Array(maxNodes * 3);
     this.opacityArray = new Float32Array(maxNodes);
 
-    // Sort indirection: _sortIndices[sortedPos] = origIdx
-    this._sortIndices = new Uint16Array(maxNodes);
+    // Sort indirection: _sortIndices[sortedPos] = origIdx.
+    // Uint32: values are node indices up to maxNodes (100k default) — Uint16
+    // wrapped past 65535, duplicating indices, which drew some nodes twice,
+    // dropped others, and made instanceId picking resolve to the wrong node.
+    this._sortIndices = new Uint32Array(maxNodes);
     this._distances = new Float32Array(maxNodes);
 
     // Camera cache — skip sort when camera hasn't moved
